% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pygradethis_setup.R
\name{add_global_setup}
\alias{add_global_setup}
\title{Helper function to execute global setup code for learnr tutorials}
\usage{
add_global_setup(code)
}
\arguments{
\item{code}{a character}
}
\value{
nothing
}
\description{
Helper function to execute global setup code for learnr tutorials
}
