% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{check_columns}
\alias{check_columns}
\title{Checks that the columns of two DataFrame/Series are the same.}
\usage{
check_columns(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{A DataFrame/Series containing the expected column.}
}
\value{
A TRUE if equal, FALSE otherwise
}
\description{
This extracts columns from the two objects, converts them to R lists 
and checks that they are identical.
}
\examples{
# ADD_EXAMPLES_HERE
}
