% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{check_index}
\alias{check_index}
\title{Checks that the Index of two DataFrame/Series are the same.}
\usage{
check_index(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{A DataFrame/Series containing the expected column.}
}
\value{
A TRUE if equal, FALSE otherwise
}
\description{
Extracts the values out of the Index to do the check.
}
\examples{
# ADD_EXAMPLES_HERE
}
