% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{check_py_dataframe}
\alias{check_py_dataframe}
\title{Checks that two DataFrame/Series are the same.}
\usage{
check_py_dataframe(
  object = .result,
  expected = .solution,
  env = parent.frame()
)
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{The expected DataFrame/Series.}
}
\value{
A TRUE if equal, FALSE otherwise
}
\description{
This will do the following checks:
1. Check that the Index are equal
2. Check that the columns are equal
3. Check that the values are equal
}
\examples{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
.solution <- reticulate::py_eval("pd.DataFrame({'a':[1,2]})")
check_py_dataframe()

.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
.solution <- reticulate::py_eval("pd.DataFrame({'b':[1,2,3]})")
check_py_dataframe()

# TODO add a MultIindex column e.g.
}
