% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{check_values}
\alias{check_values}
\title{Checks that the values of two DataFrame/Series are the same.}
\usage{
check_values(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{A DataFrame/Series to be compared to `expected`.}
}
\value{
A TRUE if equal, FALSE otherwise
}
\description{
This extracts values from the two objects, converts them to R lists
and checks that they are identical.
}
\examples{
# ADD_EXAMPLES_HERE
}
