% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_exercise_feedback}
\alias{evaluate_exercise_feedback}
\title{A helper function to test evaluation of a mocked Python exercise through learnr}
\usage{
evaluate_exercise_feedback(ex, envir = NULL, evaluate_global_setup = FALSE)
}
\arguments{
\item{ex}{A mocked Python exercise}

\item{envir}{An environment}

\item{evaluate_global_setup}{Whether to evaluate global setup code}
}
\value{
The feedback list
}
\description{
This is an internal function used for testing purposes.
}
