% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{find_arguments_mod}
\alias{find_arguments_mod}
\title{The Python module `pygradecode.find_arguments` containing functions
to check for arguments within function calls.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
find_arguments_mod
}
\value{
the Python module pygradecode.find_arguments
}
\description{
The Python module `pygradecode.find_arguments` containing functions
to check for arguments within function calls.
}
\keyword{datasets}
