% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{find_attributes_mod}
\alias{find_attributes_mod}
\title{The Python module `pygradecode.find_attributes` containing functions
to check for attributes, properties, and method chains.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
find_attributes_mod
}
\value{
the Python module pygradecode.find_attributes
}
\description{
The Python module `pygradecode.find_attributes` containing functions
to check for attributes, properties, and method chains.
}
\keyword{datasets}
