% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{find_functions}
\alias{find_functions}
\title{Get the module containing find functions for functions.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
find_functions
}
\value{
the Python module pygradecode.find_functions
}
\description{
This will provide access to particular flavors of find_*() functions
to get function and method calls.
}
\keyword{datasets}
