% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{find_functions_mod}
\alias{find_functions_mod}
\title{The `pygradecode.find_functions` module containing functions to
check for function calls.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
find_functions_mod
}
\value{
the Python module pygradecode.find_functions
}
\description{
The `pygradecode.find_functions` module containing functions to
check for function calls.
}
\keyword{datasets}
