% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_friendly_classes.R
\name{friendly_class,ANY-method}
\alias{friendly_class,ANY-method}
\alias{friendly_class,py_tbl_df-method}
\alias{friendly_class,py_grouped_df-method}
\alias{friendly_class,py_int-method}
\alias{friendly_class,py_float-method}
\alias{friendly_class,py_complex-method}
\alias{friendly_class,py_bool-method}
\alias{friendly_class,py_str-method}
\alias{friendly_class,py_tuple-method}
\alias{friendly_class,py_dict-method}
\alias{friendly_class,py_set-method}
\title{Generate a human-readable description of an object's class}
\usage{
\S4method{friendly_class}{ANY}(object)

\S4method{friendly_class}{py_tbl_df}(object)

\S4method{friendly_class}{py_grouped_df}(object)

\S4method{friendly_class}{py_int}(object)

\S4method{friendly_class}{py_float}(object)

\S4method{friendly_class}{py_complex}(object)

\S4method{friendly_class}{py_bool}(object)

\S4method{friendly_class}{py_str}(object)

\S4method{friendly_class}{py_tuple}(object)

\S4method{friendly_class}{py_dict}(object)

\S4method{friendly_class}{py_set}(object)
}
\arguments{
\item{object}{An object whose class will be described}
}
\value{
A \link{character} string of length 1,
based on the \link{class} and \link{length} of \code{object}.
}
\description{
Generate a human-readable description of an object's class
}
