% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{get_columns}
\alias{get_columns}
\title{Get the columns of a pandas.DataFrame/pandas.Series}
\usage{
get_columns(data)
}
\arguments{
\item{data}{A DataFrame/Series.}
}
\value{
An np.array of values
}
\description{
Get the columns of a pandas.DataFrame/pandas.Series
}
\examples{
# ADD_EXAMPLES_HERE
}
