% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_envir_diff}
\alias{get_envir_diff}
\title{Find the difference of variables created in prepped versus student environment}
\usage{
get_envir_diff(...)
}
\value{
the list of key strings
}
\description{
This is used to find new variables introduced in `.py_envir_result`
excluding the `.py_envir_prep` setup environment
}
\keyword{internal}
