% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_friendly_class}
\alias{get_friendly_class}
\title{Get a friendly single string representation of a Python
object's class}
\usage{
get_friendly_class(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
character
}
\description{
Get a friendly single string representation of a Python
object's class
}
