% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{get_index}
\alias{get_index}
\title{Get the index of a pandas.DataFrame/pandas.Series}
\usage{
get_index(data)
}
\arguments{
\item{data}{A DataFrame/Series.}
}
\value{
An Inex
}
\description{
Get the index of a pandas.DataFrame/pandas.Series
}
\examples{
# ADD_EXAMPLES_HERE
}
