% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_last_value}
\alias{get_last_value}
\title{Get the last value of Python source code}
\usage{
get_last_value(...)
}
\value{
the value
}
\description{
Get the last value of Python source code
}
\keyword{internal}
