% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_py_envir}
\alias{get_py_envir}
\title{Convert variables within a Python module/environment.}
\usage{
get_py_envir(envir)
}
\arguments{
\item{envir}{A Python dictionary.}
}
\value{
A named list
}
\description{
This takes in a Python dictionary representing the Python
environment, converts each object (if possible) and returns
a named list of R objects.
}
