% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{get_python_result}
\alias{get_python_result}
\title{Get the Python result}
\usage{
get_python_result(object = .result, env = parent.frame())
}
\arguments{
\item{object}{The .result object.}

\item{env}{The environment to look for the .result.}
}
\value{
A Python object
}
\description{
Get the Python result
}
\keyword{internal}
