% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{get_python_solution}
\alias{get_python_solution}
\title{Get the Python solution}
\usage{
get_python_solution(expected = .solution, env = parent.frame())
}
\arguments{
\item{expected}{The .solution object.}

\item{env}{The environment to look for the .solution.}
}
\value{
A Python object
}
\description{
Get the Python solution
}
\keyword{internal}
