% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{get_values}
\alias{get_values}
\title{Get the values of a pandas.DataFrame/pandas.Series}
\usage{
get_values(data)
}
\arguments{
\item{data}{A DataFrame/Series.}
}
\value{
An np.array of values
}
\description{
Get the values of a pandas.DataFrame/pandas.Series
}
\examples{
# ADD_EXAMPLES_HERE
}
