% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_hinted_class_messages.R
\name{hinted_class_message,py_tbl_df,py_grouped_df-method}
\alias{hinted_class_message,py_tbl_df,py_grouped_df-method}
\alias{hinted_class_message,py_grouped_df,py_tbl_df-method}
\alias{hinted_class_message,py_NotSet,ANY-method}
\title{Generate a hint for how to convert one object type to another}
\usage{
\S4method{hinted_class_message}{py_tbl_df,py_grouped_df}(object, expected)

\S4method{hinted_class_message}{py_grouped_df,py_tbl_df}(object, expected)

\S4method{hinted_class_message}{py_NotSet,ANY}(object, expected)
}
\arguments{
\item{object}{An object to be compared to \code{expected}}

\item{expected}{An object of the expected class}
}
\value{
A \link{character} string of length 1
}
\description{
Generate a hint for how to convert one object type to another
}
