% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_to_list}
\alias{index_to_list}
\title{Helper function to unpack an Index and return the values}
\usage{
index_to_list(obj)
}
\arguments{
\item{obj}{A type of Index}
}
\value{
A list()
}
\description{
Helper function to unpack an Index and return the values
}
