% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_CategoricalIndex}
\alias{is_CategoricalIndex}
\title{Checks if the Python object is a pandas.CategoricalIndex}
\usage{
is_CategoricalIndex(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a pandas.CategoricalIndex
}
