% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_DataFrame}
\alias{is_DataFrame}
\title{Checks if the Python object is a pandas.DataFrame}
\usage{
is_DataFrame(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a pandas.DataFrame
}
