% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_Index}
\alias{is_Index}
\title{Checks if the Python object is a pandas.Index}
\usage{
is_Index(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a pandas.Index
}
