% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_None}
\alias{is_None}
\title{Checks if the Python object is a None}
\usage{
is_None(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a None
}
