% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numpy_array}
\alias{is_numpy_array}
\title{Checks if the Python object is a numpy.array}
\usage{
is_numpy_array(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a numpy.array
}
