% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_RangeIndex}
\alias{is_RangeIndex}
\title{Checks if the Python object is a function}
\usage{
is_RangeIndex(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a function
}
