% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_Series}
\alias{is_Series}
\title{Checks if the Python object is a pandas.Series}
\usage{
is_Series(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a pandas.Series
}
