% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_set}
\alias{is_set}
\title{Checks if the Python object is a set}
\usage{
is_set(obj)
}
\arguments{
\item{obj}{Python object.}
}
\value{
TRUE if so, FALSE otherwise
}
\description{
Checks if the Python object is a set
}
