% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mock_py_exercise}
\alias{mock_py_exercise}
\title{A helper function to mock a Python exercise in learnr.}
\usage{
mock_py_exercise(user_code, solution_code, check, ...)
}
\arguments{
\item{user_code}{Python code submitted by the user}

\item{solution_code}{Code provided within the "-solution" chunk for the
exercise.}

\item{check}{Code provided within the "-check" chunk for the exercise.}

\item{...}{Extra grading arguments}
}
\value{
The mocked exercise
}
\description{
This is an internal function used for testing purposes.
}
