% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pygradethis_setup.R
\name{prepare_py}
\alias{prepare_py}
\title{Helper function to execute Python global setup code for learnr tutorials}
\usage{
prepare_py(code)
}
\arguments{
\item{code}{a character}
}
\value{
nothing
}
\description{
Helper function to execute Python global setup code for learnr tutorials
}
