% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem}
\alias{problem}
\title{Declare a problem}
\usage{
problem(type, expected = NULL, actual = NULL, correct = FALSE, ...)
}
\arguments{
\item{type}{A character string, e.g. `columns`, `index`, or `values` that
describes the problem that was discovered.}

\item{expected}{actual The expected and actual values. These should be
included when the value is a summary, e.g. `nrow(expected)` or
`length(actual)`. Be careful not to include large amounts of data.}

\item{...}{Additional elements to be included in the `problem` object.}
}
\description{
This is a shim around `tblcheck::problem` for constructing a `pygradethis_problem`
to communicate the problem that was discovered during checking.
}
\keyword{internal}
