% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem_grade.pygradethis_problem}
\alias{problem_grade.pygradethis_problem}
\title{Generic problem with the code}
\usage{
\method{problem_grade}{pygradethis_problem}(problem, max_diffs = 3, env = parent.frame(), ...)
}
\description{
Generic problem with the code
}
