% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem_message.pygradecode_problem}
\alias{problem_message.pygradecode_problem}
\title{Generic problem with the code}
\usage{
\method{problem_message}{pygradecode_problem}(problem, ...)
}
\description{
Generic problem with the code
}
\keyword{internal}
