% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_arguments.R
\name{py_args}
\alias{py_args}
\title{Wrapper for `pygradecode.find_arguments.args()`}
\usage{
py_args(...)
}
\arguments{
\item{...}{Any number of (un)named arguments}
}
\description{
In Python, the args() function takes positional (*args)
and keyword (**kwargs). Similarly, `py_args()` can also take
arbitrary number of positional and named arguments.
}
\details{
NOTE: the value of arguments still have to be passed as `character()`
}
