% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_check_dataframe}
\alias{py_check_dataframe}
\title{Checks that two DataFrame are the same.}
\usage{
py_check_dataframe(
  object = .result,
  expected = .solution,
  env = parent.frame()
)
}
\arguments{
\item{object}{A DataFrame to be compared to `expected`.}

\item{expected}{The expected DataFrame.}
}
\value{
A NULL if equal, otherwise a `gradethis_graded` or `problem` object
}
\description{
If DataFrames are tibble-able, rely on tblcheck::tbl_check() feedback
}
\details{
Otherwise the following checks are performed:
1. Check that the Index are equal (see `pygradethis::py_check_index()`)
2. Check that the columns are equal (see `pygradethis::py_check_columns()`)
3. Check that the values are equal (see `pygradethis::py_check_values()`)
}
