% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_check_values}
\alias{py_check_values}
\title{Checks that the values of two DataFrame/Series are the same.}
\usage{
py_check_values(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{The expected DataFrame/Series}

\item{env}{The environment used for grading.}
}
\value{
A TRUE if equal, FALSE otherwise
}
\description{
This extracts values from the two objects, converts them to R lists
and checks that they are identical.
}
\examples{
\dontrun{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2,99]})")
.solution <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
py_check_values() # FALSE
}
}
