% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_arguments.R
\name{py_find_arguments}
\alias{py_find_arguments}
\title{Wrapper for `pygradecode.find_arguments.find_arguments()`}
\usage{
py_find_arguments(code = .user_code, match = py_args(), env = parent.frame())
}
\description{
Wrapper for `pygradecode.find_arguments.find_arguments()`
}
