% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_functions.R
\name{py_find_lambdas}
\alias{py_find_lambdas}
\title{Wrapper for `pygradecode.find_functions.find_lambdas()`}
\usage{
py_find_lambdas(code = .user_code, env = parent.frame())
}
\arguments{
\item{code}{A `character` representing the code text.}

\item{env}{The environment which you don't have to worry for practical uses.}
}
\value{
A Python list[Element]
See `help('pygradecode.find_arguments.find_lambads')` in Python for more details
}
\description{
Wrapper for `pygradecode.find_functions.find_lambdas()`
}
