% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_attributes.R
\name{py_find_properties}
\alias{py_find_properties}
\title{Wrapper for `pygradecode.find_attributes.find_properties()`}
\usage{
py_find_properties(
  code = .user_code,
  match = character(1),
  env = parent.frame()
)
}
\arguments{
\item{code}{A character of the user code}

\item{match}{A character() of the property name, "" by default}

\item{env}{The grading environment}
}
\description{
Wrapper for `pygradecode.find_attributes.find_properties()`
}
