% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_get_index}
\alias{py_get_index}
\title{Get the index of a pandas.DataFrame/pandas.Series}
\usage{
py_get_index(data)
}
\arguments{
\item{data}{A DataFrame/Series.}
}
\value{
An Index/MultiIndex
}
\description{
Get the index of a pandas.DataFrame/pandas.Series
}
\examples{
\dontrun{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
pygradethis::py_get_index(.result) 

# RangeIndex(start=0, stop=3, step=1)
}
}
