% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_get_values}
\alias{py_get_values}
\title{Get the values of a pandas.DataFrame/pandas.Series}
\usage{
py_get_values(data)
}
\arguments{
\item{data}{A DataFrame/Series.}
}
\value{
An np.array of values
}
\description{
Get the values of a pandas.DataFrame/pandas.Series
}
\examples{
\dontrun{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
py_get_values(.result)

#      [,1]
# [1,]    1
# [2,]    2
# [3,]    3
}
}
