% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_grade_index}
\alias{py_grade_index}
\title{Check the index of a DataFrame/Series matches the expected index,
and return a `gradethis::fail` if they don't.}
\usage{
py_grade_index(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A DataFrame/Series to be compared to `expected`.}

\item{expected}{A DataFrame/Series containing the expected index.}

\item{env}{The environment used for grading.}
}
\value{
A `gradethis::fail()` if the index does not match else NULL
}
\description{
Check the index of a DataFrame/Series matches the expected index,
and return a `gradethis::fail` if they don't.
}
\examples{
\dontrun{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result <- reticulate::py_eval("pd.DataFrame({'a':[1,2]})")
.solution <- reticulate::py_eval("pd.DataFrame({'a':[1,2,3]})")
py_grade_index() # a gradethis::fail()
}
}
