% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{py_grade_series}
\alias{py_grade_series}
\title{Checks that two Series are the same and returns
a `gradethis::fail` if they don't.}
\usage{
py_grade_series(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{A Series to be compared to `expected`.}

\item{expected}{The expected Series}

\item{env}{The environment used for grading.}
}
\value{
A `gradethis::fail()` if the index does not match else NULL
}
\description{
Checks that two Series are the same and returns
a `gradethis::fail` if they don't.
}
\examples{
\dontrun{
reticulate::py_run_string('import pandas as pd; import numpy as np')
.result = reticulate::py_eval("pd.Series(data=[1, 2])", F)
.solution = reticulate::py_eval("pd.Series(data=[1, 2, 3])", F)
pygradethis::py_grade_series()

# <gradethis_graded: [Incorrect]
# The Series do not match the expected Series. Your result should
# contain 3 values, but it has 2 values. I expected your result to
# include the value `3`.
}
}
