% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_checker.R
\name{py_gradethis_exercise_checker}
\alias{py_gradethis_exercise_checker}
\title{A shim around the `gradethis::py_gradethis_exercise_checker` for grading Python exercises.}
\usage{
py_gradethis_exercise_checker(
  label = NULL,
  solution_code = NULL,
  user_code = NULL,
  check_code = NULL,
  envir_result = NULL,
  evaluate_result = NULL,
  envir_prep = NULL,
  last_value = NULL,
  ...
)
}
\arguments{
\item{label}{Label for exercise chunk}

\item{solution_code}{Code provided within the “-solution” chunk for the
exercise.}

\item{user_code}{Python code submitted by the user}

\item{check_code}{Code provided within the “-check” chunk for the exercise.}

\item{envir_result}{The R environment after the execution of the chunk
which also contains a Python main module environment}

\item{evaluate_result}{The return value from the `evaluate::evaluate`
function.}

\item{envir_prep}{A copy of the R environment before the execution of the
chunk which also contains a Python main module environment.}

\item{last_value}{The last value from evaluating the exercise.}

\item{...}{Extra arguments supplied by learnr}
}
\value{
The `gradethis::graded()` list which contains several fields indicating the 
result of the check.
}
\description{
To enable exercise checking in your learnr tutorial through R code (i.e. an R -check chunk), 
you can set `tutorial_options(exercise.checker = pygradethis::py_gradethis_exercise_checker)` 
in the setup chunk of your tutorial. Or, set the `exercise.checker` for an individual Python chunk.
}
