% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_identical}
\alias{py_identical}
\title{A wrapper around `identical()` to compare R objects that were
converted from Python.}
\usage{
py_identical(x, y)
}
\arguments{
\item{x}{any R object}

\item{y}{any R object}
}
\value{
TRUE/FALSE
}
\description{
A wrapper around `identical()` to compare R objects that were
converted from Python.
}
