% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_arguments.R
\name{py_literal}
\alias{py_literal}
\title{A Python class used to represent a literal string}
\usage{
py_literal(source)
}
\arguments{
\item{source}{A `character` representing code}
}
\description{
A Python class used to represent a literal string
}
