% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_user_object_get.R
\name{py_solution_object_exists}
\alias{py_solution_object_exists}
\title{Function for checking if an object was created by solution code}
\usage{
py_solution_object_exists(x, check_env = parent.frame())
}
\arguments{
\item{x}{An object name, given as a quoted `character` string.}

\item{check_env}{The `environment` from which to retrieve `.py_envir_solution`}
}
\value{
logical
}
\description{
Function for checking if an object was created by solution code
}
