% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_to_r}
\alias{py_to_r}
\title{Wrapper of reticulate::py_to_r to convert objects from Python to R.}
\usage{
py_to_r(obj)
}
\arguments{
\item{obj}{A Python object.}
}
\value{
An R object, or Python object if we can't py_to_r
}
\description{
`py_to_r` will attempt to reasonably convert objects that `reticulate::py_to_r` either can't
or when it produces a messy object (e.g. DataFrames with a MultiIndex)
}
