% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_to_tbl}
\alias{py_to_tbl}
\title{Converts a Python pandas.DataFrame into an R tibble}
\usage{
py_to_tbl(data)
}
\arguments{
\item{data}{A pandas.DataFrame}
}
\value{
A tibble
}
\description{
Converts a Python pandas.DataFrame into an R tibble
}
