% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_user_object_get.R
\name{py_user_object_exists}
\alias{py_user_object_exists}
\title{Function for checking if a Python object was created by student code}
\usage{
py_user_object_exists(x, check_env = parent.frame())
}
\arguments{
\item{x}{An object name, given as a quoted `character` string.}

\item{check_env}{The `environment` from which to retrieve `.py_envir_result`}
}
\value{
logical
}
\description{
Function for checking if a Python object was created by student code
}
