% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_user_object_get.R
\name{py_user_object_get}
\alias{py_user_object_get}
\title{Function for fetching an object created by student code}
\usage{
py_user_object_get(x, py_envir, check_env = parent.frame())
}
\arguments{
\item{x}{An object name, given as a quoted `character` string.}

\item{check_env}{The `environment` from which to retrieve `.py_envir_result`}
}
\value{
a Python object
}
\description{
Function for fetching an object created by student code
}
