% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_user_object_get.R
\name{py_user_object_list}
\alias{py_user_object_list}
\title{Function for fetching a `list` of variable names (`character`) that the
student code introduces.}
\usage{
py_user_object_list(check_env = parent.frame())
}
\arguments{
\item{check_env}{The `environment` from which to retrieve `.py_envir_result`}

\item{x}{An object name, given as a quoted `character` string.}
}
\value{
a `[character]` vector
}
\description{
The list returned contain variables introduced beyond the ones in the
`.py_envir_prep` environment which is produced by the setup code.
}
