% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_find_functions.R
\name{py_uses_function}
\alias{py_uses_function}
\title{Wrapper for `pygradecode.find_functions.uses_function()`}
\usage{
py_uses_function(code = .user_code, match = character(1), env = parent.frame())
}
\arguments{
\item{code}{A `character` representing the code text.}

\item{match}{A `character` vector of function names.}

\item{env}{The environment which you don't have to worry for practical uses.}
}
\value{
A Python list[Element]
}
\description{
See `help('pygradecode.find_arguments.uses_function')` in Python for more details
}
