% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pygradecode}
\alias{pygradecode}
\title{The `pygradecode` module that contains code checking functions.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
pygradecode
}
\value{
the `pygradecode` Python module
}
\description{
The `pygradecode` module that contains code checking functions.
}
\keyword{datasets}
