% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pygradethis}
\alias{pygradethis}
\title{The main `pygradethis` module that can be used to access other
submodules and functions within the Python library.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
pygradethis
}
\value{
the `pygradethis` Python module
}
\description{
The main `pygradethis` module that can be used to access other
submodules and functions within the Python library.
}
\keyword{datasets}
