% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pygradethis_exercise_checker}
\alias{pygradethis_exercise_checker}
\title{The `exercise.checker` pygradethis_exercise_checker to use in learnr}
\usage{
pygradethis_exercise_checker(...)
}
\value{
the `pygradethis_exercise_checker` Python function
}
\description{
The `exercise.checker` pygradethis_exercise_checker to use in learnr
}
\keyword{internal}
