% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{python_grade_learnr}
\alias{python_grade_learnr}
\title{The `exercise.checker` python_grade_learnr to use in learnr}
\usage{
python_grade_learnr(...)
}
\value{
the `python_grade_learnr` Python function
}
\description{
The `exercise.checker` python_grade_learnr to use in learnr
}
