% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{return_if_problem}
\alias{return_if_problem}
\title{Helper function to return early from `py_check_*` functions}
\usage{
return_if_problem(problem, env = parent.frame())
}
\arguments{
\item{problem}{A `pygradethis_problem` object.}

\item{env}{The environment from which to return.}
}
\description{
This is especially useful when returning from checking functions that
combine multiple checks
}
\keyword{internal}
