% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate}
\alias{translate}
\title{Wrapper of reticulate::py_to_r to translate objects from Python to R.}
\usage{
translate(obj)
}
\arguments{
\item{obj}{A Python object.}
}
\value{
An R object, or Python object if we can't translate
}
\description{
Wrapper of reticulate::py_to_r to translate objects from Python to R.
}
