% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem_message.pygradecode_problem}
\alias{problem_message.pygradecode_problem}
\title{Problem with the result}
\usage{
\method{problem_message}{pygradecode_problem}(problem, ...)
}
\description{
Problem with the result
}
