% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem_message.pygradethis_problem}
\alias{problem_message.pygradethis_problem}
\title{Generic problem with the result}
\usage{
\method{problem_message}{pygradethis_problem}(problem, ...)
}
\description{
Generic problem with the result
}
\keyword{internal}
