% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_to_r}
\alias{py_to_r}
\title{Wrapper of reticulate::py_to_r to py_to_r objects from Python to R.}
\usage{
py_to_r(obj)
}
\arguments{
\item{obj}{A Python object.}
}
\value{
An R object, or Python object if we can't py_to_r
}
\description{
Wrapper of reticulate::py_to_r to py_to_r objects from Python to R.
}
